Formatações e testes unitários

+++ EasyIteris/Entities/SimulationQueryRequest.swift
@@ -10,11 +10,11 @@ import Foundation
 struct SimulationQueryRequest: Encodable {
     let investedAmount: Double
     let index: String
-    let rate: Int
+    let rate: Double
     let isTaxFree: Bool
     let maturityDate: String
 
-    init(investedAmount: Double, index: String = "CDI", rate: Int, isTaxFree: Bool = false, maturityDate: String) {
+    init(investedAmount: Double, index: String = "CDI", rate: Double, isTaxFree: Bool = false, maturityDate: String) {
         self.investedAmount = investedAmount
         self.index = index
         self.rate = rate
